pyproject.toml
[project]
name = "lls-auth-client"
version = "0.1.0"
description = "A FastAPI client with Keycloak OAuth2 for secure access to Llama Stack and MCP tools."
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "httpx==0.27.2",
    "python-dotenv==1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "pytest-asyncio==0.24.0",
    "ruff==0.6.8"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
